requirements/requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.4
Django==3.2.5
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-taggit==1.5.1
mysqlclient==2.0.3
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
